Add files via upload
I broke my code a bunch in the end, and scrambled and lost track of everything that I was doing.
I hope I can get some points for this, or the chance to rework it.  I am not expecting much out of this one.
I also overwrote the read.md file and was unable to reproduce it.